Revert to button group spacing version
Revert commits after e3c0655 to restore the app to the last known working state before the Add Items merge and follow-up receipt extraction changes.
