+++ src/pages/UploadReceipt.css
@@ -1,0 +1,130 @@
+.upload-receipt-container {
+  max-width: 440px;
+  margin: 24px auto;
+  padding: 24px;
+  background: #fff;
+  border-radius: 16px;
+  box-shadow: 0 2px 16px rgba(0, 0, 0, 0.07);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.upload-receipt-container h2 {
+  margin-bottom: 8px;
+}
+
+.upload-receipt-container p {
+  margin-bottom: 16px;
+  color: #555;
+  text-align: center;
+}
+
+.upload-btn,
+.next-btn {
+  background: #6366f1;
+  color: #fff;
+  border: none;
+  border-radius: 8px;
+  padding: 10px 24px;
+  margin: 0;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: background 0.2s;
+}
+
+.upload-btn:hover,
+.next-btn:hover:enabled {
+  background: #4f46e5;
+}
+
+.next-btn:disabled {
+  background: #c7d2fe;
+  cursor: not-allowed;
+}
+
+.image-actions {
+  width: 100%;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 10px;
+  margin-bottom: 12px;
+}
+
+.image-input {
+  display: none;
+}
+
+.selected-image-name {
+  font-size: 0.92rem;
+  margin: 6px 0;
+  color: #6b7280;
+  word-break: break-all;
+}
+
+.input-divider {
+  margin: 10px 0 8px;
+  color: #6b7280;
+  font-size: 0.92rem;
+}
+
+.json-paste-section {
+  width: 100%;
+  margin-top: 2px;
+}
+
+.json-paste-section label {
+  display: block;
+  font-weight: 600;
+  margin-bottom: 8px;
+  text-align: left;
+  color: #374151;
+}
+
+.json-paste-section textarea {
+  width: 100%;
+  box-sizing: border-box;
+  border: 1px solid #d1d5db;
+  border-radius: 8px;
+  padding: 10px;
+  font-family: monospace;
+  font-size: 0.95rem;
+  resize: vertical;
+  margin-bottom: 10px;
+}
+
+.receipt-preview {
+  width: 100%;
+  max-height: 260px;
+  object-fit: contain;
+  border: 1px solid #e5e7eb;
+  border-radius: 8px;
+  margin: 8px 0 10px;
+  background: #f9fafb;
+}
+
+.json-error,
+.json-success,
+.json-status {
+  font-size: 0.95rem;
+  margin: 6px 0 0 0;
+  text-align: center;
+}
+
+.json-error {
+  color: #b91c1c;
+}
+
+.json-success {
+  color: #047857;
+}
+
+.json-status {
+  color: #4338ca;
+}
+
+.btn-group {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
